pre_processing_code
I've changed the structure of dat long such that each row represents a single trial (rather than a single response). I've also added in the full text of each excerp,  the headline and the word count as variables. Note that this pre-processing script at the three data files it produces are all in the google drive folder as well.